requirements.txt
fastmcp==2.7.0
mcp==1.9.3
uvicorn==0.32.1
fastapi==0.115.6
pyjwt==2.10.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
httpx==0.28.1
cryptography==43.0.3